Se cambio el parseo a tipo float #352

+++ app/assets/javascripts/geomaps.js
@@ -432,7 +432,7 @@ Navarra.geomaps = function (){
       success: function(data){
         count_row = []
         $.each(data,function(a,b){
-          count_row.push(parseInt(b['count']));
+          count_row.push(parseFloat(b['count']));
         })
 
         min = Math.min(...count_row);
